添加设备&模板增删改通知到协议驱动SDK

+++ dc3-center/dc3-manager/src/main/java/com/pnoker/center/manager/service/impl/DriverInfoServiceImpl.java
@@ -20,9 +20,11 @@ import com.baomidou.mybatisplus.core.conditions.query.LambdaQueryWrapper;
 import com.baomidou.mybatisplus.core.toolkit.Wrappers;
 import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
 import com.pnoker.center.manager.mapper.DriverInfoMapper;
+import com.pnoker.center.manager.message.ManagerMessageSender;
 import com.pnoker.center.manager.service.DriverInfoService;
 import com.pnoker.common.bean.Pages;
 import com.pnoker.common.constant.Common;
+import com.pnoker.common.constant.Operation;
 import com.pnoker.common.dto.DriverInfoDto;
 import com.pnoker.common.exception.ServiceException;
 import com.pnoker.common.model.DriverInfo;
@@ -46,6 +48,8 @@ import java.util.Optional;
 public class DriverInfoServiceImpl implements DriverInfoService {
     @Resource
     private DriverInfoMapper driverInfoMapper;
+    @Resource
+    private ManagerMessageSender messageSender;
 
     @Override
     @Caching(
@@ -64,6 +68,7 @@ public class DriverInfoServiceImpl implements DriverInfoService {
             throw new ServiceException("driver info already exists");
         }
         if (driverInfoMapper.insert(driverInfo) > 0) {
+            messageSender.notifyDriver(Operation.Profile.UPDATE, driverInfo.getProfileId());
             return driverInfoMapper.selectById(driverInfo.getId());
         }
         return null;
@@ -79,7 +84,12 @@ public class DriverInfoServiceImpl implements DriverInfoService {
             }
     )
     public boolean delete(Long id) {
-        return driverInfoMapper.deleteById(id) > 0;
+        DriverInfo driverInfo = selectById(id);
+        if (null != driverInfo) {
+            messageSender.notifyDriver(Operation.Profile.UPDATE, driverInfo.getProfileId());
+            return driverInfoMapper.deleteById(id) > 0;
+        }
+        throw new ServiceException("driver info does not exist");
     }
 
     @Override
@@ -95,7 +105,12 @@ public class DriverInfoServiceImpl implements DriverInfoService {
     )
     public DriverInfo update(DriverInfo driverInfo) {
         driverInfo.setUpdateTime(null);
+        DriverInfo select = selectByDriverAttributeId(driverInfo.getDriverAttributeId(), driverInfo.getProfileId());
+        if (null != select) {
+            throw new ServiceException("driver info already exists");
+        }
         if (driverInfoMapper.updateById(driverInfo) > 0) {
+            messageSender.notifyDriver(Operation.Profile.UPDATE, driverInfo.getProfileId());
             return selectById(driverInfo.getId());
         }
         return null;
